Move Groups/Tiles/groupTiles.jsx to Groups/Tiles/index.jsx
It makes more sense for the groups page to import from Groups/Tiles, not
Groups/Tiles/groupTiles.
