+++ static-site/src/pages/groups.jsx
@@ -5,7 +5,7 @@ import * as splashStyles from "../components/splash/styles";
 import { fetchAndParseJSON } from "../util/datasetsHelpers";
 import DatasetSelect from "../components/Datasets/dataset-select";
 import ListResources from "../components/ListResources/index";
-import { GroupTiles } from "../components/Groups/Tiles/groupTiles";
+import { GroupTiles } from "../components/Groups/Tiles";
 import GenericPage from "../layouts/generic-page";
 import { UserContext } from "../layouts/userDataWrapper";
 import { DataFetchErrorParagraph } from "../components/splash/errorMessages";
